add GoodsCard
