+++ src/components/personal/goods/Goods.jsx
@@ -5,6 +5,7 @@ import './Goods.css';
 import GoodsNewBot from './GoodsNewBot/GoodsNewBot';
 import Dropdown from '../../showcase/Dropdown/Dropdown';
 import { getMyProducts } from '../../../store/actions';
+import GoodsCard from './GoodsCard/GoodsCard';
 
 const Goods = () => {
   const [isShown, setIsShown] = useState(false);
@@ -22,6 +23,11 @@ const Goods = () => {
     dispatch(getMyProducts());
     console.log(results);
   }, []);
+
+  function handleEditBot() {
+    setIsShown(true);
+    setIsShownGoods(false);
+  }
   return (
     <section className='goods'>
       <button
@@ -32,12 +38,14 @@ const Goods = () => {
       >
         Добавить новый
       </button>
-      {isShown && <GoodsNewBot />}
+      {isShown && <GoodsNewBot shownNewBot={() => setIsShown(false)} />}
       {isShownGoods && results && (
         <>
           <Dropdown />
           {results?.map((item) => {
-            return <div key={item.id}></div>;
+            return (
+              <GoodsCard key={item.id} item={item} setIsShown={handleEditBot} />
+            );
           })}
         </>
       )}
