+++ src/components/personal/goods/GoodsNewBot/GoodsNewBot.jsx
@@ -14,7 +14,7 @@ import PopupPriceBot from '../../../popups/PopupPrice/PopupPriceBot';
 import { useFormAndValid } from '../../../../hooks/useFormAndValid';
 import { postProduct } from '../../../../store/actions';
 
-const GoodsNewBot = () => {
+const GoodsNewBot = (props) => {
   const dispatch = useDispatch();
   const [showCategoryPopup, setShowCategoryPopup] = useState(false);
   const [showNamePopup, setShowNamePopup] = useState(false);
@@ -99,6 +99,7 @@ const GoodsNewBot = () => {
   function handleSubmit(e) {
     e.preventDefault();
     dispatch(postProduct(formValue));
+    props.shownNewBot();
     // Нужно добавить функцию для сброса формы или перехода на страницу со своими ботами
   }
